+++ core/dbt/task/debug.py
@@ -5,10 +5,11 @@ import sys
 from typing import Optional, Dict, Any, List
 
 from dbt.logger import GLOBAL_LOGGER as logger
+from dbt import flags
 import dbt.clients.system
 import dbt.exceptions
 from dbt.adapters.factory import get_adapter, register_adapter
-from dbt.config import Project, Profile, PROFILES_DIR
+from dbt.config import Project, Profile
 from dbt.config.renderer import DbtProjectYamlRenderer, ProfileRenderer
 from dbt.config.utils import parse_cli_vars
 from dbt.context.base import generate_base_context
@@ -69,7 +70,7 @@ class QueryCommentedProfile(Profile):
 class DebugTask(BaseTask):
     def __init__(self, args, config):
         super().__init__(args, config)
-        self.profiles_dir = getattr(self.args, 'profiles_dir', PROFILES_DIR)
+        self.profiles_dir = flags.PROFILES_DIR
         self.profile_path = os.path.join(self.profiles_dir, 'profiles.yml')
         try:
             self.project_dir = get_nearest_project_dir(self.args)
@@ -156,7 +157,7 @@ class DebugTask(BaseTask):
             self.project = Project.from_project_root(
                 self.project_dir,
                 renderer,
-                verify_version=getattr(self.args, 'version_check', False),
+                verify_version=flags.VERSION_CHECK,
             )
         except dbt.exceptions.DbtConfigError as exc:
             self.project_fail_details = str(exc)
@@ -195,7 +196,7 @@ class DebugTask(BaseTask):
             try:
                 partial = Project.partial_load(
                     os.path.dirname(self.project_path),
-                    verify_version=getattr(self.args, 'version_check', False),
+                    verify_version=bool(flags.VERSION_CHECK),
                 )
                 renderer = DbtProjectYamlRenderer(
                     generate_base_context(self.cli_vars)